LayoutTests/http/tests/security/cross-origin-blob-transfer.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
async_test(t => {
    assert_equals(window.location.origin, "http://127.0.0.1:8000");
    window.addEventListener("message", t.step_func(e => {
        e.data.text().then(t.step_func_done(text => {
            assert_equals(text, "Blob content");
        }));
    }));
}, "Test for creating blob in iframe and then transferring it cross-origin.");
</script>
<iframe src="http://localhost:8000/security/resources/iframe-cross-origin-blob-transfer.html"></iframe>
</body>
</html>
